Update .sh code for inference
I have added a feature that allows the user to choose a different singer from the configs/singers directory for each input file. The user can now select a singer from a menu or a prompt, and the corresponding spk file name will be passed to the svc_inference.py command. I have also added some validation to check if the user's input is valid and matches one of the options.